Added filtering option for depth map and some performance measurement code.
* Image scale factor configurable
* Better error handling

+++ python/RLrecon/engine/unreal_cv_wrapper.py
@@ -11,10 +11,14 @@ from RLrecon import math_utils
 
 class UnrealCVWrapper(BaseEngine):
 
-    class Exception(BaseException):
+    class Exception(RuntimeError):
         pass
 
-    def __init__(self, address=None, port=None, image_scale_factor=0.5):
+    def __init__(self, address=None, port=None,
+                 image_scale_factor=0.5,
+                 max_depth_distance=np.finfo(np.float).max,
+                 max_depth_viewing_angle=math_utils.degrees_to_radians(90.)):
+        super(UnrealCVWrapper, self).__init__(max_depth_distance, max_depth_viewing_angle)
         self._width = None
         self._height = None
         self._image_scale_factor = image_scale_factor
@@ -25,7 +29,7 @@ class UnrealCVWrapper(BaseEngine):
         self._cv_client = Client((address, port))
         self._cv_client.connect()
         if not self._cv_client.isconnected():
-            raise(Exception("Unable to connect to UnrealCV"))
+            raise(self.Exception("Unable to connect to UnrealCV"))
 
     def _scale_image(self, image, interpolation_mode=cv2.INTER_CUBIC):
         """Scale an image to the desired size"""
@@ -74,6 +78,8 @@ class UnrealCVWrapper(BaseEngine):
     def get_rgb_image(self):
         """Return the current RGB image"""
         img_str = self._cv_client.request('vget /camera/0/lit png')
+        if img_str is None:
+            raise self.Exception("UnrealCV request failed")
         img = np.fromstring(img_str, np.uint8)
         rgb_image = cv2.imdecode(img, cv2.IMREAD_COLOR)
         rgb_image = self._scale_image(rgb_image)
@@ -82,30 +88,61 @@ class UnrealCVWrapper(BaseEngine):
     def get_rgb_image_by_file(self):
         """Return the current RGB image (transport via filesystem)"""
         filename = self._cv_client.request('vget /camera/0/lit lit.png')
+        if filename is None:
+            raise self.Exception("UnrealCV request failed")
         rgb_image = cv2.imread(filename)
         rgb_image = self._scale_image(rgb_image)
         os.remove(filename)
         return rgb_image
 
-    def get_normal_image(self):
-        """Return the current normal image"""
+    def get_normal_rgb_image(self):
+        """Return the current normal image in RGB encoding (i.e. 128 is 0, 0 is -1, 255 is +1)"""
         img_str = self._cv_client.request('vget /camera/0/normal png')
+        if img_str is None:
+            raise self.Exception("UnrealCV request failed")
         img = np.fromstring(img_str, np.uint8)
         normal_image = cv2.imdecode(img, cv2.IMREAD_COLOR)
         normal_image = self._scale_image_with_nearest_interpolation(normal_image)
         return normal_image
 
-    def get_normal_image_by_file(self):
-        """Return the current normal image (transport via filesystem)"""
+    def get_normal_rgb_image_by_file(self):
+        """Return the current normal image in RGB encoding (i.e. 128 is 0, 0 is -1, 255 is +1)
+        (transport via filesystem)
+        """
         filename = self._cv_client.request('vget /camera/0/normal normal.png')
+        if filename is None:
+            raise self.Exception("UnrealCV request failed")
         normal_image = cv2.imread(filename)
         normal_image = self._scale_image_with_nearest_interpolation(normal_image)
         os.remove(filename)
         return normal_image
 
+    def get_normal_image(self):
+        """Return the current normal image in vector representation"""
+        normal_rgb_image = self.get_normal_rgb_image()
+        # Image is in BGR ordering, so [z, y, x]. Convert to RGB.
+        normal_rgb_image = normal_rgb_image[:, :, ::-1]
+        normal_image = self._convert_normal_rgb_image_to_normal_image(normal_rgb_image, inplace=False)
+        # Convert left-handed Unreal system to right-handed system
+        normal_image[:, :, 1] = -normal_image[:, :, 1]
+        self.filter_normal_image(normal_image)
+        return normal_image
+
+    def get_normal_image_by_file(self):
+        """Return the current normal image in vector representation (transport via filesystem)"""
+        normal_rgb_image = self.get_normal_rgb_image_by_file()
+        # Image is in BGR ordering, so [z, y, x]. Convert to RGB.
+        normal_rgb_image = normal_rgb_image[:, :, ::-1]
+        normal_image = self._convert_normal_rgb_image_to_normal_image(normal_rgb_image, inplace=False)
+        # Convert left-handed Unreal system to right-handed system
+        normal_image[:, :, 1] = -normal_image[:, :, 1]
+        return normal_image
+
     def get_ray_distance_image(self):
         """Return the current ray-distance image"""
         img_str = self._cv_client.request('vget /camera/0/depth npy')
+        if img_str is None:
+            raise self.Exception("UnrealCV request failed")
         img_str_io = StringIO(img_str)
         ray_distance_image = np.load(img_str_io)
         ray_distance_image = self._scale_image_with_nearest_interpolation(ray_distance_image)
@@ -114,6 +151,8 @@ class UnrealCVWrapper(BaseEngine):
     def get_ray_distance_image_by_file(self):
         """Return the current ray-distance image (transport via filesystem)"""
         filename = self._cv_client.request('vget /camera/0/depth depth.exr')
+        if filename is None:
+            raise self.Exception("UnrealCV request failed")
         ray_distance_image = cv2.imread(filename, cv2.IMREAD_ANYDEPTH)
         ray_distance_image = self._scale_image_with_nearest_interpolation(ray_distance_image)
         os.remove(filename)
@@ -136,6 +175,8 @@ class UnrealCVWrapper(BaseEngine):
     def get_location(self):
         """Return the current location in meters as [x, y, z]"""
         location_str = self._cv_client.request('vget /camera/0/location')
+        if location_str is None:
+            raise self.Exception("UnrealCV request failed")
         location_unreal = np.array([float(v) for v in location_str.split()])
         # Convert location from Unreal (cm) to meters
         location_unreal *= 0.01
@@ -146,6 +187,8 @@ class UnrealCVWrapper(BaseEngine):
     def get_orientation_rpy(self):
         """Return the current orientation in radians as [roll, pitch, yaw]"""
         orientation_str = self._cv_client.request('vget /camera/0/rotation')
+        if orientation_str is None:
+            raise self.Exception("UnrealCV request failed")
         pitch, yaw, roll = [float(v) * np.pi / 180. for v in orientation_str.split()]
         euler_rpy = np.array([roll, pitch, yaw])
         roll, pitch, yaw = math_utils.convert_rpy_from_left_to_right_handed(euler_rpy)
@@ -186,7 +229,9 @@ class UnrealCVWrapper(BaseEngine):
         request_str = 'vset /camera/0/location {:f} {:f} {:f}'.format(
             location_unreal[0], location_unreal[1], location_unreal[2])
         # print("Sending location request: {}".format(request_str))
-        self._cv_client.request(request_str)
+        response = self._cv_client.request(request_str)
+        if response != "ok":
+            raise self.Exception("UnrealCV request failed: {}".format(response))
 
     def set_orientation_rpy(self, roll, pitch, yaw):
         """Set new orientation in radians"""
@@ -195,7 +240,9 @@ class UnrealCVWrapper(BaseEngine):
         request_str = 'vset /camera/0/rotation {:f} {:f} {:f}'.format(
             pitch * 180 / np.pi, yaw * 180 / np.pi, roll * 180 / np.pi)
         # print("Sending orientation request: {}".format(request_str))
-        self._cv_client.request(request_str)
+        response = self._cv_client.request(request_str)
+        if response != "ok":
+            raise self.Exception("UnrealCV request failed: {}".format(response))
 
     def set_orientation_quat(self, quat):
         """Set new orientation quaterion quat = [w, x, y, z]"""
